feat(cli): add success output to silent mutation commands
Show short IDs in task link output instead of full 32-char IDs.
Add SuccessMessage with page count to generate man-pages command.
Tag/untag commands already have success output from the store refactor.

+++ src/cli/commands/task/link.rs
@@ -60,7 +60,7 @@ impl Command {
       store::write_task(config, &target_task)?;
     }
 
-    let msg = format!("Linked {} --{}--\u{003e} {}", id, self.rel, target_id);
+    let msg = format!("Linked {} --{}--\u{003e} {}", id.short(), self.rel, target_id.short());
     println!("{}", SuccessMessage::new(&msg, theme));
     Ok(())
   }
